test(ImagePreview): sort test cases
